modificando dto

+++ tarjetacredito/src/main/java/edu/um/ar/programacion2/tarjetacredito/repository/TarjetaCreditoRepository.java
@@ -8,7 +8,7 @@ import edu.um.ar.programacion2.tarjetacredito.model.TarjetaCredito;
 import edu.um.ar.programacion2.tarjetacredito.objeto.TarjetaCreditoObjeto;
 
 public interface TarjetaCreditoRepository extends JpaRepository<TarjetaCredito, Long> {
-	Optional<TarjetaCredito> findByNumero(Integer numero);
+	Optional<TarjetaCredito> findByNumero(Long numero);
 	Boolean existsByNumero(Long numero);
 	Boolean existsByToken(String token);
 	Optional<TarjetaCredito> findByToken(String token);
